requirements.txt
python-dotenv==1.0.0
flask==3.0.0
markdown==3.5.1
google-generativeai==0.3.2
python-csv==0.0.13
